chore: move shared capabilities to capabilities.rs
